refactor(lotto): 출력 메서드를 OutputHandler로 이동
